pyproject.toml
[tool.poetry]
name = "atomic-habbits"
version = "0.1.0"
description = ""
authors = ["exetch <gassemark@gmail.com>"]
readme = "README.md"
packages = [{include = "atomic_habbits"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
pillow = "9.5.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.4"
faker = "^20.1.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
